0.2.9 - Tracker.py - Removed PythonExtension for http arguments instead. Added variables to be added into the /tracker URL. Replaced some redirects to url_for to ensure that the variables are sent to the /tracker url. Added a logout feature to make sure that you session isn't saved in your browser
tracker.html - Removed old feature dealing with sessions. Instead, using toasts that are picked up from the error arguments passed to it by other functions. Will allow me to not use {% py %} and make it look a little cleaner to read as well. Fixed text saying "test" to now display their squad position. Also made it load any toasts that are on the webpage to display errors in the bottom right of the browser.

PythonExtension.py - Removed and replaced by toasts from Bootstrap